refactor: convert pages to React Server Components (RSC)
- Convert news page from client to server component with dynamic metadata
- Convert creator page from client to server component with generateMetadata
- Extract client-side functionality to dedicated *-page-content.tsx components
- Create modular creator components: CreatorSummary, CreatorAnalyses, CreatorComments
- Implement server/client component delegation pattern
- Add comprehensive RSC refactoring documentation
- Improve SEO with dynamic metadata generation
- Enhance performance with server-side rendering
- Maintain all interactive functionality in client components
- Follow Next.js 13+ App Router best practices

Benefits:
- Better SEO with server-rendered metadata
- Improved Core Web Vitals scores
- Reduced client bundle size
- Faster initial page loads
- Better separation of concerns
