+++ app/news/page.tsx
@@ -1,36 +1,41 @@
 /**
- * News Page - Main page for displaying trending analyses and creator sources
+ * News Page - Server component for displaying trending analyses and creator sources
  *
  * This page shows:
  * - A feed of all analyses (main content)
  * - Top credible sources (sidebar)
  * - Top misinformation sources (sidebar)
+ *
+ * This is now a React Server Component that renders a client component
+ * for interactivity while enabling better performance and SEO.
  */
 
-"use client";
+import { Metadata } from "next";
+import { NewsPageContent } from "@/components/news";
 
-import { AllAnalyses } from "@/components/all-analyses";
-import { useLanguage } from "@/components/language-provider";
-import {
-  TopCredibleSources,
-  TopMisinformationSources,
-} from "@/components/news";
+/**
+ * Metadata for the news page
+ */
+export const metadata: Metadata = {
+  title: "Trending News Analysis | Checkmate",
+  description:
+    "Discover trending content analyses, credible sources, and fact-checked information from across social media platforms.",
+  keywords: [
+    "news analysis",
+    "fact checking",
+    "content verification",
+    "social media",
+    "credibility",
+  ],
+};
 
+/**
+ * News page server component
+ *
+ * This server component provides better SEO, faster initial page loads,
+ * and enables server-side rendering while delegating client interactions
+ * to the NewsPageContent component.
+ */
 export default function NewsPage() {
-  const { t } = useLanguage();
-  return (
-    <div className="grid grid-cols-1 md:grid-cols-3 gap-6 max-w-7xl mx-auto p-4">
-      {/* Main Content Feed */}
-      <div className="md:col-span-2 space-y-6">
-        <h1 className="text-2xl font-bold">{t.trendingOnCheckmate}</h1>
-        <AllAnalyses />
-      </div>
-
-      {/* Right Sidebar */}
-      <div className="space-y-6 hidden md:block sticky top-4 self-start max-h-[calc(100vh-2rem)] overflow-y-auto pt-15">
-        <TopCredibleSources />
-        <TopMisinformationSources />
-      </div>
-    </div>
-  );
+  return <NewsPageContent />;
 }
