+++ components/creator/index.ts
@@ -1,0 +1,14 @@
+/**
+ * Creator Components Module
+ *
+ * This module exports all creator-related components for easy importing
+ * throughout the application.
+ */
+
+export { CreatorSummary } from "./creator-summary";
+export { CreatorAnalyses } from "./creator-analyses";
+export { CreatorComments } from "./creator-comments";
+export { CreatorPageContent } from "./creator-page-content";
+
+// Re-export types
+export type { Creator } from "./creator-summary";
